refactor: group related react components
