+++ components/profile/UserCollections.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../components/styles/Profile.module.css";
+import styles from "../styles/Profile.module.css";
 
 export default function UserCollections({
   collections,
